+++ spec/rubocop/cop/style/redundant_regexp_escape_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe RuboCop::Cop::Style::RedundantRegexpEscape do
     end
 
     [
-      ('a'..'z').to_a - %w[c n p u x],
+      ('a'..'z').to_a - %w[c g k n p u x],
       ('A'..'Z').to_a - %w[C M P],
       %w[n101 x41 u0041 u{0041} cc C-c p{alpha} P{alpha}]
     ].flatten.each do |escape|
@@ -43,6 +43,14 @@ RSpec.describe RuboCop::Cop::Style::RedundantRegexpEscape do
         it 'does not register an offense' do
           expect_no_offenses("foo = /[\\#{escape}]/")
         end
+      end
+    end
+
+    context "with an invalid \g escape" do
+      it 'does not register an offense' do
+        # See https://ruby-doc.org/core-2.7.1/Regexp.html#class-Regexp-label-Subexpression+Calls
+        # \g should be \g<name>
+        expect_no_offenses('foo = /\g/')
       end
     end
 
@@ -78,6 +86,19 @@ RSpec.describe RuboCop::Cop::Style::RedundantRegexpEscape do
         end
       end
     end
+
+      context "with an escaped '+' inside a character class inside a group" do
+        it 'registers an offense and corrects' do
+          expect_offense(<<~'RUBY')
+            foo = /([\+])/
+                     ^^ Redundant escape inside regexp literal
+          RUBY
+
+          expect_correction(<<~RUBY)
+            foo = /([+])/
+          RUBY
+        end
+      end
 
     context 'with an escaped . inside a character class beginning with :' do
       it 'registers an offense and corrects' do